services/dashboard/requirements.txt
aiohttp
aiohttp_jinja2
aiohttp-devtools
pyyaml
motor